refactor(runtime): simplify shared query code
Separate session ownership checks and put query entry points before their helpers.

Share OpenCode identity reads and remove redundant names, types, and comments.
