+++ packages/host/src/adapters/agent-sessions/runtime-query-adapter.ts
@@ -13,6 +13,37 @@ import type {
   NativeAgentRuntimeQueries,
 } from "../../ports/agent-runtime-query-port";
 
+export const createRuntimeQueryAdapter = (
+  native: NativeAgentRuntimeQueries & AgentSessionQueryParentPort,
+): AgentRuntimeQueryAdapterPort => {
+  const read = <Input extends RuntimeQueryIdentity, Result>(
+    operation: string,
+    input: Input,
+    run: () => Promise<Result>,
+  ) =>
+    Effect.tryPromise({ try: run, catch: (cause) => toRuntimeQueryError(operation, input, cause) });
+  return {
+    resolveSessionParent: (input) =>
+      read("read session parent", input, () => native.resolveSessionParent(input)),
+    listAvailableModels: (input) =>
+      read("list models", input, () => native.listAvailableModels(input)),
+    listAvailableSlashCommands: (input) =>
+      read("list slash commands", input, () => native.listAvailableSlashCommands(input)),
+    listAvailableSkills: (input) =>
+      read("list skills", input, () => native.listAvailableSkills(input)),
+    listAvailableSubagents: (input) =>
+      read("list subagents", input, () => native.listAvailableSubagents(input)),
+    searchFiles: (input) => read("search files", input, () => native.searchFiles(input)),
+    loadSessionHistory: (input) =>
+      read("load session history", input, () => native.loadSessionHistory(input)),
+    loadSessionTodos: (input) =>
+      read("load session todos", input, () => native.loadSessionTodos(input)),
+    loadSessionDiff: (input) =>
+      read("load session diff", input, () => native.loadSessionDiff(input)),
+    loadFileStatus: (input) => read("load file status", input, () => native.loadFileStatus(input)),
+  };
+};
+
 export const toRuntimeQueryError = (
   operation: string,
   input: RuntimeQueryIdentity,
@@ -48,35 +79,4 @@ export const toRuntimeQueryError = (
       : "The runtime query failed. Check the host runtime logs and retry this read.",
     cause,
   );
-};
-
-export const createRuntimeQueryAdapter = (
-  native: NativeAgentRuntimeQueries & AgentSessionQueryParentPort,
-): AgentRuntimeQueryAdapterPort => {
-  const read = <Input extends RuntimeQueryIdentity, Result>(
-    operation: string,
-    input: Input,
-    run: () => Promise<Result>,
-  ) =>
-    Effect.tryPromise({ try: run, catch: (cause) => toRuntimeQueryError(operation, input, cause) });
-  return {
-    resolveSessionParent: (input) =>
-      read("read session parent", input, () => native.resolveSessionParent(input)),
-    listAvailableModels: (input) =>
-      read("list models", input, () => native.listAvailableModels(input)),
-    listAvailableSlashCommands: (input) =>
-      read("list slash commands", input, () => native.listAvailableSlashCommands(input)),
-    listAvailableSkills: (input) =>
-      read("list skills", input, () => native.listAvailableSkills(input)),
-    listAvailableSubagents: (input) =>
-      read("list subagents", input, () => native.listAvailableSubagents(input)),
-    searchFiles: (input) => read("search files", input, () => native.searchFiles(input)),
-    loadSessionHistory: (input) =>
-      read("load session history", input, () => native.loadSessionHistory(input)),
-    loadSessionTodos: (input) =>
-      read("load session todos", input, () => native.loadSessionTodos(input)),
-    loadSessionDiff: (input) =>
-      read("load session diff", input, () => native.loadSessionDiff(input)),
-    loadFileStatus: (input) => read("load file status", input, () => native.loadFileStatus(input)),
-  };
 };
