+++ packages/host/src/interface/commands/agent-runtime-query-command-handlers.ts
@@ -15,6 +15,47 @@ import type { AgentRuntimeQueryPort } from "../../ports/agent-runtime-query-port
 import type { HostCommandHandlerDefinitions } from "../router/host-command-router";
 import type { HostCommandArgs } from "./command-inputs";
 
+export const createAgentRuntimeQueryCommandHandlers = (service: AgentRuntimeQueryPort) =>
+  ({
+    agent_runtime_list_models: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listModels,
+      (input) => service.listAvailableModels(input),
+      (input, result) => result.runtime === undefined || result.runtime.kind === input.runtimeKind,
+    ),
+    agent_runtime_list_slash_commands: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSlashCommands,
+      (input) => service.listAvailableSlashCommands(input),
+    ),
+    agent_runtime_list_skills: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSkills,
+      (input) => service.listAvailableSkills(input),
+    ),
+    agent_runtime_list_subagents: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSubagents,
+      (input) => service.listAvailableSubagents(input),
+    ),
+    agent_runtime_search_files: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.searchFiles,
+      (input) => service.searchFiles(input),
+    ),
+    agent_runtime_load_session_history: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionHistory,
+      (input) => service.loadSessionHistory(input),
+    ),
+    agent_runtime_load_session_todos: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionTodos,
+      (input) => service.loadSessionTodos(input),
+    ),
+    agent_runtime_load_session_diff: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionDiff,
+      (input) => service.loadSessionDiff(input),
+    ),
+    agent_runtime_file_status: createQueryHandler(
+      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.fileStatus,
+      (input) => service.loadFileStatus(input),
+    ),
+  }) satisfies HostCommandHandlerDefinitions;
+
 const envelopeSchema = z.object({ input: z.unknown() }).strict();
 const identitySchema = z.object({
   repoPath: z.string().optional().catch(undefined),
@@ -66,44 +107,3 @@ const createQueryHandler =
       }
       return result;
     });
-
-export const createAgentRuntimeQueryCommandHandlers = (service: AgentRuntimeQueryPort) =>
-  ({
-    agent_runtime_list_models: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listModels,
-      (input) => service.listAvailableModels(input),
-      (input, result) => result.runtime === undefined || result.runtime.kind === input.runtimeKind,
-    ),
-    agent_runtime_list_slash_commands: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSlashCommands,
-      (input) => service.listAvailableSlashCommands(input),
-    ),
-    agent_runtime_list_skills: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSkills,
-      (input) => service.listAvailableSkills(input),
-    ),
-    agent_runtime_list_subagents: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.listSubagents,
-      (input) => service.listAvailableSubagents(input),
-    ),
-    agent_runtime_search_files: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.searchFiles,
-      (input) => service.searchFiles(input),
-    ),
-    agent_runtime_load_session_history: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionHistory,
-      (input) => service.loadSessionHistory(input),
-    ),
-    agent_runtime_load_session_todos: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionTodos,
-      (input) => service.loadSessionTodos(input),
-    ),
-    agent_runtime_load_session_diff: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.loadSessionDiff,
-      (input) => service.loadSessionDiff(input),
-    ),
-    agent_runtime_file_status: createQueryHandler(
-      AGENT_RUNTIME_QUERY_COMMAND_CONTRACTS.fileStatus,
-      (input) => service.loadFileStatus(input),
-    ),
-  }) satisfies HostCommandHandlerDefinitions;
